# src/api-docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "fafifu-befe-fp",
    "contact": {}
  },
  "host": "fafifu-backend-api.herokuapp.com",
  "basePath": "/v1",
  "securityDefinitions": {
    "apiKey": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "schemes": [
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/product/wishlist": {
      "get": {
        "summary": "Wishlist",
        "tags": [
          "Wishlist"
        ],
        "operationId": "Wishlist",
        "deprecated": false,
        "produces": [
          "application/json; charset=utf-8"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "examples": {
              "application/json; charset=utf-8": {
                "data": [
                  {
                    "publicId": "0225dec5-33a1-467f-8567-592628c2ec5f",
                    "name": "Chicken",
                    "description": "New range of formal shirts are designed keeping you in mind. With fits and styling that will make you stand apart",
                    "price": 286,
                    "category": [
                      {
                        "categoryId": 4,
                        "name": "Elektronik"
                      },
                      {
                        "categoryId": 4,
                        "name": "Elektronik"
                      },
                      {
                        "categoryId": 1,
                        "name": "Hobi"
                      }
                    ],
                    "imageUrl": [
                      {
                        "imageUrl": "http://idealistic-walker.info"
                      },
                      {
                        "imageUrl": "https://unruly-sarong.org"
                      },
                      {
                        "imageUrl": "http://fearless-arcade.net"
                      },
                      {
                        "imageUrl": "https://downright-member.biz"
                      },
                      {
                        "imageUrl": "https://needy-veteran.net"
                      }
                    ]
                  }
                ]
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "examples": {
              "application/json; charset=utf-8": {
                "message": "Unauthorized User"
              }
            }
          },
          "404": {
            "description": "Not Found",
            "examples": {
              "application/json; charset=utf-8": {
                "message": "Wishlist list not found"
              }
            }
          }
        },
        "security": [
          {
            "apiKey": []
          }
        ]
      }
    },
    "/product/{id}/wishlist": {
      "post": {
        "summary": "Add Wishlist",
        "tags": [
          "Wishlist"
        ],
        "operationId": "AddWishlist",
        "deprecated": false,
        "produces": [
          "application/json; charset=utf-8"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "examples": {
              "application/json; charset=utf-8": {
                "message": "Success add wishlist"
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "examples": {
              "application/json; charset=utf-8": {
                "message": "Product already in wishlist"
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "examples": {
              "application/json; charset=utf-8": {
                "message": "Unauthorized User"
              }
            }
          }
        },
        "security": [
          {
            "apiKey": []
          }
        ]
      },
      "delete": {
        "summary": "Delete Wishlist",
        "tags": [
          "Wishlist"
        ],
        "operationId": "DeleteWishlist",
        "deprecated": false,
        "produces": [
          "application/json; charset=utf-8"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "examples": {
              "application/json; charset=utf-8": {
                "message": "Success delete wishlist"
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "examples": {
              "application/json; charset=utf-8": {
                "message": "Unauthorized User"
              }
            }
          },
          "404": {
            "description": "Not Found",
            "examples": {
              "application/json; charset=utf-8": {
                "message": "Wishlist not found"
              }
            }
          }
        },
        "security": [
          {
            "apiKey": []
          }
        ]
      }
    },
...
    "/notification/{id}/remove": {
      "delete": {
        "summary": "Remove Notification",
        "tags": [
          "Notification"
        ],
        "operationId": "RemoveNotification",
        "deprecated": false,
        "produces": [
          "application/json; charset=utf-8"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "examples": {
              "application/json; charset=utf-8": {
                "message": "Success remove notification"
              }
            }
          },
          "404": {
            "description": "Not Found",
            "examples": {
              "application/json; charset=utf-8": {
                "message": "Notification not found"
              }
            }
          }
        },
        "security": [
          {
            "apiKey": []
          }
        ]
      }
    },
    "/notification/remove-all": {
      "delete": {
        "summary": "Remove All Notification",
        "tags": [
          "Notification"
        ],
        "operationId": "RemoveAllNotification",
        "deprecated": false,
        "produces": [
          "application/json; charset=utf-8"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "examples": {
              "application/json; charset=utf-8": {
                "message": "Success remove all notification"
              }
            }
          },
          "404": {
            "description": "Not Found",
            "examples": {
              "application/json; charset=utf-8": {
                "message": "Notification list not found"
              }
            }
          }
        },
        "security": [
          {
            "apiKey": []
          }
        ]
      }
    }
  },
  "security": [],
...
}